+++ asyncutils/iotools.py
@@ -15,12 +15,12 @@ double_ended_text_pipe, double_ended_binary_pipe = t = tuple(map(f, ('r', 'rb'),
 P.patch_function_signatures(*((_, s) for _ in t))
 @H.subscriptable
 class AsyncReadWriteCouple(A.LoopContextMixin):
-    __slots__ = '_aso', '_executor', 'reader', 'writer'; executor, AMBIGUOUS_CALL_TIP = None, 'tip: delegate to reader or writer as appropriate'
+    __slots__ = '__aso', '__x', 'reader', 'writer'; executor = None
     def __init__(self, r, w, /, executor=None, _=H.create_executor, *, find_attr_on_writer_first=False):
         if not r.readable(): raise TypeError(f'asyncutils.iotools.AsyncReadWriteCouple: reader {r!r} is not readable')
         if not w.writable(): raise TypeError(f'asyncutils.iotools.AsyncReadWriteCouple: writer {w!r} is not writable')
-        super().__init__(); self._executor, self.reader, self.writer, self._aso = _(t) if executor is None is (executor := (t := type(self)).executor) else executor, r, w, (w, r) if find_attr_on_writer_first else (r, w)
-    async def _run(self, f, *a): return await self.loop.run_in_executor(self._executor, f, *a)
+        super().__init__(); self.__x, self.reader, self.writer, self.__aso = _(t) if executor is None is (executor := (t := type(self)).executor) else executor, r, w, (w, r) if find_attr_on_writer_first else (r, w)
+    async def _run(self, f, *a): return await self.loop.run_in_executor(self.__x, f, *a)
     def read(self, n=-1, /): return self._run(self.reader.read, n)
     def read1(self, n=-1, /): return self._run(self.reader.read1, n)
     def readall(self): return self._run(r.read if (f := getattr(r := self.reader, 'readall', None)) is None else f)
@@ -34,21 +34,23 @@ class AsyncReadWriteCouple(A.LoopContextMixin):
     def readlines(self, hint=-1, /): return self._run(self.reader.readlines, hint)
     def write(self, s, /): return self._run(self.writer.write, s)
     def writelines(self, lines, /): return self._run(self.writer.writelines, lines)
-    def fileno(self): A.raise_exc(OSError, E.EBADF, 'asyncutils.iotools.AsyncReadWriteCouple: ambiguous file descriptor query', notes=self.AMBIGUOUS_CALL_TIP)
-    def isatty(self): A.raise_exc(OSError, E.ENOTSUP, 'asyncutils.iotools.AsyncReadWriteCouple: ambiguous isatty query', notes=self.AMBIGUOUS_CALL_TIP)
+    @staticmethod
+    def __rx(e, m, t='tip: delegate to reader or writer as appropriate', /): A.raise_exc(OSError, e, f'asyncutils.iotools.AsyncReadWriteCouple: ambiguous {m} call', notes=t)
+    def fileno(self): self.__rx(E.EBADF, 'fileno')
+    def isatty(self): self.__rx(E.ENOTSUP, 'isatty')
     readable = writable = lambda _, /: True
     def flush(self): return self._run(self.writer.flush)
     def seekable(self): return False # noqa: PLR6301
-    def seek(self, offset, whence=0, /): A.raise_exc(OSError, E.ESPIPE, 'asyncutils.iotools.AsyncReadWriteCouple: cannot use seek on read-write couple', notes=self.AMBIGUOUS_CALL_TIP) # noqa: ARG002
-    def tell(self): A.raise_exc(OSError, E.ESPIPE, 'asyncutils.iotools.AsyncReadWriteCouple: tell is not supported', notes=self.AMBIGUOUS_CALL_TIP)
+    def seek(self, offset, whence=0, /): self.__rx(E.ESPIPE, 'seek') # noqa: ARG002
+    def tell(self): self.__rx(E.ESPIPE, 'tell')
     def truncate(self, size=None, /): return self._run(self.writer.truncate, size)
-    async def aclose(self): await gather(*map(self._run, (self.reader.close, self.writer.close))); self._executor.shutdown()
+    async def aclose(self): await gather(*map(self._run, (self.reader.close, self.writer.close))); self.__x.shutdown()
     __cleanup__ = aclose
     @property
     def closed(self): return self.reader.closed and self.writer.closed
     def __getattr__(self, n, /):
         f = (a := []).append
-        for _ in self._aso:
+        for _ in self.__aso:
             try: return getattr(_, n)
             except AttributeError as e: f(e)
         raise ExceptionGroup(f'asyncutils.iotools.AsyncReadWriteCouple: did not find attribute {n!r}', a) from None