+++ tests/test_init.py
@@ -1,8 +1,8 @@
-import asyncutils
+import asyncutils as au
 def test_mod():
-    assert isinstance(asyncutils.__getattr__, type)
-    assert isinstance(asyncutils.submodules_map, dict)
-    assert asyncutils.preloaded_submodules.issuperset(('constants', 'context', 'cli', 'exceptions', 'version'))
-    assert asyncutils.console_preloaded_submodules.issuperset(asyncutils.preloaded_submodules)
-    assert asyncutils.time_since_boot() > 0
-    assert asyncutils.VersionInfo(asyncutils.__hexversion__) == asyncutils.__version__
+    assert isinstance(au.__getattr__, type)
+    assert isinstance(au.submodules_map, dict)
+    assert au.preloaded_submodules.issuperset(('constants', 'context', 'cli', 'exceptions', 'version'))
+    assert au.console_preloaded_submodules.issuperset(au.preloaded_submodules)
+    assert au.time_since_boot() > 0
+    assert au.VersionInfo(au.__hexversion__) == au.__version__